cleanup, add new example files
